scheduler-app/backend/requirements.txt
flask
flask-cors
pandas
numpy
ortools
supabase
python-dotenv
bcrypt
PyJWT